Add initial class for test method analyzer
Add skeleton for analyzer of test methods, meaning capturing
behaviour that is the same for both TestCaseAttribute and
TestAttribute, like the ExpectedResult property.
Currently, there is some duplication between TestMethodUsageAnalyzer
and TestCaseUsageAnalyzer.

Also, no tests have been added for tests marked with TestAttribute,
but the existing tests for TestCaseAttribute have been moved to the
TestMethodUsage folders.