Give all analyzers unique identifiers
In order to use Gu.Roslyn.Asserts to make tests more succient, and
also to reduce the "magic" in the tests, we need that all analyzers
have unique identifers.

This PR just makes all the identifers unique, in a subsequent PR
we'll make use of Gu.Roslyn.Asserts.

See GuOrg/Gu.Roslyn.Asserts#99 for a little context.