Use RegisterCompilationStartAction to avoid repeated calls to GetTypeByMetadataName (#548)
* Use RegisterCompilationStartAction to avoid repeated calls to GetTypeByMetadataName

I noticed a number of places where there are repeated calls to GetTypeByMetadataName as the first thing an analyzer does, but not in a CompilationStartAction. By restructuring these calls, we can run that call just once for the entire compilation, avoiding repeated lookups and avoiding doing any more work on the entire compilation if those types aren't found. The compilation object does maintain a small cache of string->symbol for recent lookups, but it's best not to rely on this cache. The cache was put in as a crutch to help particularly egregious analyzers, and it's not hard to blow the limit on it. Additionally, the cache can't get skipping entire swaths of actions benefit that using CompilationStartAction can.

There's more opportunity to make this better; for example, nearly every analyzer derived from BaseAssertionAnalyzer could take advantage of this. That is a significantly more involved refactoring though, and I didn't have time to take that on at this point. This would also benefit from using the WellKnownTypesProvider in Microsoft.CodeAnalysis.AnalyzerUtilities to manage caching, and avoid comparing types by name.

---------

Co-authored-by: Manfred Brands <manfred-brands@users.noreply.github.com>